Improve time complexity
Credit goes to John Hughes for discovering this.

Data.Vector.cons is O(m) for an m-element vector (according to the
docs), while (:) is O(1). In this case, cons is used inside a loop over
all n points given to the algorithm, so that loop is something like
O(nm) instead of O(n).

More specifically, m in this case is roughly n/k, at least when the
points are distributed somewhat evenly among the k clusters. So the
inner loop has a complexity of roughly O(n²/k) instead of O(n).

In addition to being slower than necessary, this implementation also
results in very surprising results when comparing the parallel version
to the sequential one: For example, I saw a speedup of a factor 7 on 4
cores for small values of k.

This happens because the parallel version splits the points into
partitions before sending them to the loop. So in the parallel version,
the loop only has to deal with smaller sets of points, making its
quadratic behavior less impactful overall.

One should also note that increasing k limits the size of the individual
(m-element) vectors, mitigating the loop's quadratic behavior.