Fix partitions bug
From the docs:

"This version of k-means takes an additional arguments -- the number of
partitions the set of points'll be divided into. This needn't equal the
number of processors: [...]"

However, the partitions parameter is actually implemented as the number
of points per cluster – not the number of clusters. For example,
partitions == 4 means that 1000 points are divided into 250 chunks to be
processed in parallel.

This commit changes the implementation so that partitions == 4 means
that 1000 points are divided into 4 chunks instead.